Move SimulatorFullNodeService type alias to start_simulator (#17875)
Move SimulatorFullNodeService type alias to start_simulator.
