+++ chia/_tests/cmds/test_farm_cmd.py
@@ -11,7 +11,8 @@ from chia.cmds.farm_funcs import summary
 from chia.farmer.farmer import Farmer
 from chia.harvester.harvester import Harvester
 from chia.simulator.block_tools import BlockTools
-from chia.types.aliases import FarmerService, HarvesterService, SimulatorFullNodeService, WalletService
+from chia.simulator.start_simulator import SimulatorFullNodeService
+from chia.types.aliases import FarmerService, HarvesterService, WalletService
 
 
 @pytest.mark.anyio
