+++ chia/_tests/util/setup_nodes.py
@@ -36,14 +36,8 @@ from chia.simulator.setup_services import (
     setup_wallet_node,
 )
 from chia.simulator.socket import find_available_listen_port
-from chia.types.aliases import (
-    FarmerService,
-    FullNodeService,
-    HarvesterService,
-    SimulatorFullNodeService,
-    TimelordService,
-    WalletService,
-)
+from chia.simulator.start_simulator import SimulatorFullNodeService
+from chia.types.aliases import FarmerService, FullNodeService, HarvesterService, TimelordService, WalletService
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.peer_info import UnresolvedPeerInfo
 from chia.util.hash import std_hash
